Final configuration so integration tests are working
