+++ prototypes/createaccount
@@ -12,8 +12,3 @@ UPDATE_USER_REQUEST='{ "email": "after@example.com", "bio": "after a long day of
 UPDATE_USER_RESPONSE=$(curl -s -d "$UPDATE_USER_REQUEST" -X PUT -H "$AUTH_HEADER" -H "$JSON" $URL/api/v1/user)
 USER_RESP=$(curl -s -H "$AUTH_HEADER" $URL/api/v1/user)
 RECOMMENDED_USERS=$(curl -s -H "$AUTH_HEADER" $URL/api/v1/recommends)
-for i in `echo $RECOMMENDED_USERS | python3 -c $'import sys, json;\nfor x in json.load(sys.stdin):\n    print(x[\'userId\'])'`
-do
-  LIKE_USER='{ "userid": "'$i'" }'
-  curl -s -d "$LIKE_USER" -X PUT -H "$AUTH_HEADER" -H "$JSON" $URL/api/v1/like
-done
